Add disk removal alert (#134)
In a case that a disk fail and is not recognized anymore by the
kernel, the metrics of a specific device will stop from been
collected.

Virtual disks don't have the model, serial and wwn, so those
fields are used to just target physical ones. If a physical disk
is removed, the alert will continue to fire as long as the disk
remains missing and up to 30 days from the removal.